requirements.txt
mcp[cli]>=1.2.0
python-dotenv>=1.0.1
mistralai>=1.2.0
fastmcp==2.12.0
requests>=2.31.0
httpx>=0.27.0
wandb>=0.16.0
weave>=0.50.0
flask
# LangGraph and LangChain dependencies
langgraph>=0.6.7
langchain-openai>=0.2.0
langchain-core>=0.3.0
# MCP integration
langchain-mcp-adapters
